Add update kwarg to QuantumScript.copy (#6285)
**Context:**
We often make new, modified tapes in transforms, and occasionally an
attribute of the old tape is not included. It can be unclear in some
cases whether this was intentional or not. For example, in many
transforms, we don't copy over `trainable_params` when making the new
tape. It can be unclear whether this is because we are intending to set
`trainable_params` to the default `None`, or whether it's an oversight.
We've also had at least one transform where we missed copying `shots`
over until later.

**Description of the Change:**
We add an `update` kwarg to the `tape.copy` method that takes a dict
with possible keys `operations`, `measurements`, `shots` and/or
`trainable_params`. Passing a dict to `update` changes the defined
attributes while copying everything else over from the original tape.

**Benefits:**
If we use this in transforms, we won't be able to forget to copy over
tape attributes - and if we overwrite anything, it will be clearly
intentional (i.e. we will have to explicitly pass `None` to the `update`
method instead of leaving the attribute out when creating a new tape).

[[sc-65487](https://app.shortcut.com/xanaduai/story/65487)]

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: Christina Lee <christina@xanadu.ai>