`qml.wire.Wires` accepts `JAX` arrays (#6312)
**Context:** In JAX JIT/plxpr/qjit workflows, it is quite easy for wires
to end up with jax array values. JAX arrays are not hashable and are
rejected. Still, we can treat them as iterable and ensure they are
accepted as wires.

**Description of the Change:** We use the `to_list` and `item` methods
of `JAX` arrays as a trick to make them tuples before hashing them.

**Benefits:** `qml.wire.Wires` can accept `JAX` arrays as input.

**Possible Drawbacks:** The change should only be relevant when
`jax.numpy.array`s are explicitly provided. Therefore, in principle, we
should not face problems with the rest of the code.

**Related GitHub Issues:** None.

**Related Shortcut Stories** [sc-72593] [sc-74904]