Add generator checks and differentiability checks to assert_valid (#6282)
The purpose of this story is to check if every operator can be
differentiated properly with the parameter shift method. The gradient
calculated using parameter shift is compared with backprop to verify
that they are equal.

There are some cases where this does not apply:
1. Backprop does not produce the correct gradient in some cases such as
for `StatePrep` and for `QubitUnitary`, as it does not take into account
the constraint that the matrix must remain unitary, making this test
invalid.
2. Some operator takes integers as parameters, such as `BasisState`. In
this case, it does not make sense to take the gradient of integer
parameters.

For these cases, a `skip_differentiation` toggle is added to
`assert_valid` such that the differentiation check is skipped for these
operators.

Three bugs are found as a result of adding this check. The relevant
tests are xfailed:
https://github.com/PennyLaneAI/pennylane/issues/6331
https://github.com/PennyLaneAI/pennylane/issues/6333
https://github.com/PennyLaneAI/pennylane/issues/6340

Some other minor bug fixes are also included in this PR.

[sc-65197]
Fixes https://github.com/PennyLaneAI/pennylane/issues/6311