Making arithmetic templates QJIT compatible (#6307)
This PR makes the following template QJIT compatible.

1. PhaseAdder
2. Adder
3. Multiplier
4. ModExp
5. OutAdder
6. OutMultiplier

The issue stems from the use of `qml.adjoint()` as an iterable in the
PhaseAdder and Multiplier templates. In Catalyst, the corresponding
`HybridAdjoint` class is compatible with PennyLane's `Adjoint` class for
single operations, but incompatible for multiple operations (see [this
Catalyst PR](https://github.com/PennyLaneAI/catalyst/pull/768)). The
solution is add each `qml.adjoint(op)` to the decomposition list
individually, instead of extending the list by treating the adjoint as
an iterable.

While this PR makes OutAdder and OutMultiplier QJIT compatible, there is
another issue with them that prevents them from returning the correct
answer when compiled with QJIT. This bug will be addressed in a separate
PR.

This PR addresses sc-74445.