Remove `AllWires` validation and associated tests (#6373)
**Context:**

Pennylane checks every operator in the initial tape such that if
`op.num_wires == AllWires and op.wires and device.wires`, then `op.wires
== device.wires`.

This is an edge case that kinda only applies to GlobalPhase. And it
really is fine if GlobalPhase has wires that are only part of the device
wires. We kinda ignore global phase wires anyway.

**Description of the Change:**

Removes code block for validation and accompanying tests. 

**Benefits:** Prevents future performance issues and deals with
`GlobalPhase` edge case.

**Possible Drawbacks:** None

[sc-72066]