Make `qml.QutritBasisStatePreparation` JIT compatible (#6308)
This PR makes `qml.QutritBasisStatePreparation` JIT compatible.
Previously, the template used the `qml.TShift` operator in the
decomposition. However, this approach requires non-jittable control flow
on the input. To make the template jittable the decomposition is changed
to return a `qml.ops.QutritUnitary` representing a `TShift` operator
raised to some power. This change requires adding `matrix_power` and
`eigh` to the multi dispatch. This PR addresses sc-70863.

---------

Co-authored-by: Guillermo Alonso-Linaje <65235481+KetpuntoG@users.noreply.github.com>