Rename `test_simplify_with_adjoint_not_defined` in `test_pow_ops.py` (#6388)
This PR renames `test_simplify_with_adjoint_not_defined` to
`test_simplify_with_pow_not_defined` in the `test_pow_ops.py` file. The
test is given by the following code.
```
    def test_simplify_with_adjoint_not_defined(self):
        """Test the simplify method with an operator that has not defined the op.pow method."""
        op = Pow(qml.U2(1, 1, 0), z=3)
        simplified_op = op.simplify()
        assert isinstance(simplified_op, Pow)
        assert op.data == simplified_op.data
        assert op.wires == simplified_op.wires
        assert op.arithmetic_depth == simplified_op.arithmetic_depth
```
Note the function name says `adjoint` is not defined, but the docstring
says that `pow` is not defined. The docstring is correct, since `qml.U2`
defines the `adjoint` method.