Add `qml.workflow.construct_tape(qnode, level)(*args, **kwargs)` (#6419)
**Context:**

In an effort to clean-up the code in `QNode`, we need a way to construct
tapes from qnodes. This could both be used by the user and internally in
the class.

**Description of the Change:**

Create a thin wrapper around the existing `qml.workflow.construct_batch`
that returns a single tape. An error will be raised if the `level`
argument specified corresponds to more than one tape.

Example

```python
@qml.qnode(qml.device("default.qubit", shots=10))
def circuit(x):
    qml.RandomLayers(qml.numpy.array([[1.0, 2.0]]), wires=(0,1))
    qml.RX(x, wires=0)
    qml.RX(-x, wires=0)
    qml.SWAP((0,1))
    qml.X(0)
    qml.X(0)
    return qml.expval(qml.X(0) + qml.Y(0))

>>> tape = qml.workflow.construct_tape(circuit)(0.5)
>>> tape.circuit
[RandomLayers(tensor([[1., 2.]], requires_grad=True), wires=[0, 1]),
 RX(0.5, wires=[0]),
 RX(-0.5, wires=[0]),
```

**Benefits:** Code clean-up - allows us to phase out the mutable
`QNode.tape` and `QNode.qtape` properties!

**Possible Drawbacks:** None

[sc-72155]