Add `get_best_diff_method` and `_get_gradient_fn` to `qml.workflow` (#6399)
**Context:**

Right now, `QNode.get_gradient_fn` and `QNode.get_best_method`, and
`QNode.best_method_str` all exist as public or static methods and aren't
used internally (aside from inside the `QNode` class) and they don't
make much sense externally. To address this, a new user-friendly
function should be added.

**Description of the Change:**

Implemented functions `get_best_diff_method` and `_get_gradient_fn` to
`qml.workflow`. They behave the same as the existing methods in `QNode`
but are designed with a simpler interface in mind 😄. The former is
intended to be more user-facing and so has a simplifed interface. The
latter is more for internal development and will be used to deprecate
the internal `QNode` methods mentioned earlier.

For `get_best_diff_method` the user can provide the `QNode` they are
trying to execute and it will return a string with the "best"
differentiation method.

For `_get_gradient_fn`, the intention was to extract this from `QNode`
and use it later on once we begin the deprecation and removal of those
methods outlined earlier.4

Example 

_`qml.workflow.get_best_diff_method`_

```python
>>> dev = qml.device("default.qubit")
>>> qn = qml.QNode(lambda: None, dev)
>>> qml.workflow.get_best_diff_method(qn)()
'backprop'
>>> dev_shots = qml.device("default.qubit", shots=45)
>>> qn_shots = qml.QNode(lambda: None, dev_shots)
>>> qml.workflow.get_best_diff_method(qn_shots)()
'parameter-shift'
```

_`qml.workflow._get_gradient_fn`_

```python
>>> qml.workflow._get_gradient_fn(dev_shots, diff_method='parameter-shift')
<transform: param_shift>
```

**Benefits:** Improves `QNode` organization and structure. 

**Possible Drawbacks:** None

[sc-72157]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>