Use `pytest-rng` generated seeds for stochastic tests (#6435)
**Context:**

The PennyLane CI suffers from sporadic failures due to stochasticity of
test cases. In order to save CI runtime, we have seeded many tests.
However, this approach introduces the risk of a test case passing due to
a special seed, potentially hiding bugs that could have been discovered.
Therefore, it is beneficial to periodically update the seeds to review
such tests.

**Description of the Change:**

- The test suite uses `pytest-rng` to handle seed generation. A `seed`
fixture is now available to use for every test case. If seeding is
required, please use the seed provided by `pytest-rng`

```python
def test_some_function(..., seed):
    dev = qml.device("default.qubit", seed=seed)
```

- A `local_salt` test marker has been added that allows you to modify
the seed locally for a particular test case if the generated seed
happens to make your test fail.

```python
@pytest.mark.local_salt(42)
def test_some_function(..., seed):
    dev = qml.device("default.qubit", seed=seed)
```

- All tests that uses local seeds has been updated to use the new
seeding approach.

- An `rng_salt` has been added to `pytest.ini`. This controls the seed
generation for the entire test suite. We want to periodically change
this `rng_salt` which updates all the seeds across the test suite.

- The fixture that sets the global seed for every test case has been
removed, and another fixture is added that restores the global seed
after each test in case the test modifies the global seed.

**Benefits:**

We will be able to catch bugs hidden in tests that are passing only
because of a magic seed that was set locally.

**Possible Drawbacks:**

This is a chore to maintain this seed an ensure that all tests pass with
the new salt when it is updated.

**Related Shortcut Story:**

[sc-74294]

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>