[BUGFIX] `shadow_expval` doesn't work with parameter broadcasting (#6464)
**Context:** WIth `default.qubit`, parameter broadcasting does not work
with `shadow_expval` (or `classical_shadow`) as the measurement process.

These latest developments might be relevant for historical reasons:

- In [this PR](https://github.com/PennyLaneAI/pennylane/pull/4238),
support for parameter broadcasting was added for `measure` and
`measure_with_samples` (presumably as an effort for the new device API).
Specifically, a new keyword argument `is_state_batched` has been
introduced to these functions to deal with the parameter broadcasting
case.

- In [this PR](https://github.com/PennyLaneAI/pennylane/pull/4162), the
classical shadow measurement process was integrated into the new device
API as well. A new logical branch was integrated into
`measure_with_samples` adding a new `_measure_classical_shadow
function`. This function did not receive the `is_state_batched
argument`, so apparently supporting parameter broadcasting for classical
shadows was not part of the requirements (the shortcut story was not
linked so I cannot investigate further).

- In [this PR](https://github.com/PennyLaneAI/pennylane/pull/4429), the
`is_state_batched` argument has been added to
`_measure_classical_shadow`, but support for parameter broadcasting was
never implemented (the argument is unused).

**Description of the Change:** We added an internal conditional quantum
transform to `default.qubit` such that the
`qml.transforms.broadcast_expand` is conditionally applied if
`ShadowExpvalMP` or `ClassicalShadowMP` are found in the tape
measurement process.

**Benefits:** Now `default.qubit` supports (non-native) parameter
broadcasting with shadow measurement processes.

**Possible Drawbacks:** Since we are applying the `broadcast_expand`
quantum transform, we are not supporting *native* parameter
broadcasting, This should not be a big deal.

**Related GitHub Issues:** #6301 

**Related ShortCut Stories:** [sc-74387]