Adding string representation for some operators (#6493)
**Context:** The string representation format of some operators is
slightly different from others. This is because some do not have their
own string representation. Furthermore, the output of the Hadamard
operator should be updated since its string representation recently
changed.

**Description of the Change:**

- Change the output in the doc of the `qml.Hadamard` operator, since its
string representation changed during this release
- Add a string representation for the `qml.S` operator and change the
output in the doc
- Add a string representation for the `qml.T` operator and change the
output in the doc
- Add a string representation for the `qml.SX` operator and change the
output in the doc

**Benefits:** Correct doc output and more consistency among string
representation for non-parametrized single wire operators.

**Possible Drawbacks:** None that I can think of. This may cause
failures in some plugin tests that check explicitly the string
representation of such operators, but this failure should be immediate
to fix.

**Related GitHub Issues:** None.