update neighbour_order error in Lattice (#6497)
Documentation previously said `neighbour_order` was limited to "not
being more than 1" when `custom_edges` are defined. The phrasing is
potentially confusing, as it begs the question what inputs aside from 1
are allowed; physically it doesn't make sense for it to be _less_ than
one.

In practice, it was treating any input `neighbour_order < 1` as
`neighbour_order = 1`, and raising an error for any input above 1. In
https://github.com/PennyLaneAI/pennylane/pull/6481 we changed the
docstring to explicitly say "must be 1" rather than "must not be more
than 1". Here we also update the condition for raising the error, and
the error message, for consistency with the docstring.

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>