Add helper func `apply_operation` for default mixed (#6379)
**Context:**
Before getting into the implementation of new `default_mixed` we need to
separate most of the useful logics from the older version one. One of
the most important things is the `apply_operation`. Basically, what we
want to do is to condensate everything into a new sub-module under
`qml/devices/qubit_mixed` such that previous empirical heuristics
regarding how to select between `einsum` and `tensordot` can be carried
over and wrapped independently.

**Description of the Change:**
New script containing the helper funcs as described above.

**Benefits:**
Better modularization. Getting ready for the implementaion of a new
`default_mixed` in the future.

**Possible Drawbacks:**

**Related GitHub Issues:**

**Related Shortcut Stories**
[sc-73319]

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Christina Lee <christina@xanadu.ai>