Change Torch wheel install variants for CI (#6533)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [ ] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:** The existing CI infra uses Torch wheels pulled directly
from the PyTorch hosted builds. This has defaulted to pulling in the
ROCm build of the wheels, which are 2GB to download, and 12+GB when
installed and unpacked. As Github action only provide 14GB guaranteed
storage for runners, this immediately eats into the available storage
and often causes out of space failures.

**Description of the Change:** This changes the Torch wheel installs to
the default ones from PyPI. The default wheels still pull in a large
number of external deps (several GB, including Nvidia runtime libs), but
as of now allows us to remain below the 14GB threshold.

**Benefits:** Prevent out of space errors in the CI.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: albi3ro <chrissie.c.l@gmail.com>