Adding string representation for some operators (#6542)
**PLEASE READ BEFORE MERGING**

Immediately after this PR is merged into PL, [this
one](https://github.com/PennyLaneAI/catalyst/pull/1292) should be marked
as ready for review and merged into Catalyst to fix the tests.

**Context:** The string representation format of some operators is
slightly different from others. This is because some do not have their
own string representation. Furthermore, the output of the Hadamard
operator should be updated since its string representation recently
changed.

**Description of the Change:**

- Change the output in the doc of the `qml.Hadamard` operator, since its
string representation changed during the previous release
- Add a string representation for the `qml.S` operator and change the
output in the doc
- Add a string representation for the `qml.T` operator and change the
output in the doc
- Add a string representation for the `qml.SX` operator and change the
output in the doc

**Benefits:** Correct doc output and more consistency among string
representation for non-parametrized single wire operators.

**Possible Drawbacks:** This change was performed during the feature
freeze before the release of PL 0.39, but it had to be reverted because
of a single test in Catalyst that failed (the test uses the string
representation to verify the result). It is necessary to create a small
PR in Catalyst to fix such a test and link it to this story after the PL
PR is merged.

**Related GitHub Issues:** None.

**Related Shortcut Stories:** [sc-77632]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>