Rename `gradient_fn` to `diff_method` in `qml.execute` (#6549)
**Context:**

We use the term `diff_method` in the `QNode` and `gradient_fn` in
`qml.execute`. This can cause confusion having two different terms for
basically the same thing.

**Description of the Change:**

Deprecates the term `gradient_fn` and strictly uses `diff_method` in
`qml.execute`.

**Benefits:**

Increased consistency in terminology. Opens us to preprocessing the
`diff_method` *inside* of `qml.execute`, making it more robust and
user-friendly.

**Possible Drawbacks:**

Deprecations can always cause problems. This is an easy change, as it
can be handled by a find-replace automation, but that is always extra
work.

**Related GitHub Issues:**

[sc-77590]