[labs.dla 1] lie_closure_dense using dense matrices (#6371)
We identified that there are instances where computing commutators is
significantly more efficient using their dense matrix representation. We
thus provide a function to perform the lie closure using the dense
representation and make it an experimental pennylane.labs feature.

Computing the adjoint representation to follow. From there the
functionality is the same, its just a different path to obtain the
adjoint representation.

Note that this is a re-implementation of
[lie_closure](https://docs.pennylane.ai/en/latest/code/api/pennylane.lie_closure.html?highlight=lie_closure#pennylane.lie_closure).
For mathematical details, see its docs. @PietropaoloFrisoni

[sc-74994]

---------

Co-authored-by: Lee J. O'Riordan <lee@xanadu.ai>
Co-authored-by: Lee James O'Riordan <mlxd@users.noreply.github.com>
Co-authored-by: dwierichs <david.wierichs@xanadu.ai>
Co-authored-by: Jay Soni <jbsoni@uwaterloo.ca>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>