Add `_resolve_execution_config` helper function to `qnode.py` (#6498)
**Context:**

The goal is to streamline the process of updating the execution
configuration with non-device-specific settings. This approach
consolidates the setup and validation of the execution configuration in
one spot in the code, before calling `qml.execute` in `qnode.py`.

**Description of the Change:**

This refactor organizes the setup of the initial `config` by moving it
into a dedicated helper function. It also includes the setup and
validation code for handling the `MCMConfig`.

**Benefits:** Improves code cleaniness in `qnode.py` and allows for
better testing of internal pipeline.

**Possible Drawbacks:** None

[sc-72086]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>