move `qnode_call` to the `workflow` module and make it private (#6620)
**Context:**

As we are integrating more things with the capture project, we are
making changes across the entirety of pennylane. Previously, we stuck
most capture related code in the capture module, and then imported in
whichever relevant place we needed too.

I don't think this makes sense anymore. As we are tying together the
capture work to more parts of pennylane, the capture code should just be
located next to the code it is capturing.

**Description of the Change:**

Move `qml.capture.qnode_call` to the workflow module and make it
private.

**Benefits:**

Code for managing the workflow is located in the `workflow` module.

**Possible Drawbacks:**

This is all experimental, so we should be able to move things around
whenever we see fit. I'm not concerned about this being "breaking".

**Related GitHub Issues:**
[sc-78505