Fix `qnode_spectrum` with `interface="auto"` (#6622)
**Context:**
An example from the docstring of `qnode_spectrum` is failing because the
detection of trainable arguments fails when using Autograd via vanilla
numpy inputs and `interface="auto"` (default).

Note that this problem is caused by a bug in `qml.math.is_independent`,
but so far only surfaces in the specific context of `qnode_spectrum`,
which uses this function to check a Jacobian to be constant.

**Description of the Change:**
Add a warning to the docs that pure numpy parameters are not supported.
Raise an error if pure numpy parameters are present.
Smaller updates relating to code quality/linting.

**Benefits:**
Fixes #6593 

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
Closes #6593 
[sc-78453]