Opmath docs (#6589)
**Context:**
A number of documentation changes and especially type hint changes
already went in to master with the primary PR removing operator
arithmetic, but we left the larger documentation sections (unifying
docstring messaging surrounding
qml.Hamiltonian/qml.ops.LinearCombination, updating the guidelines page
for new operator arithmetic) for a separate PR.

**Description of the Change:**
- There is now a single docstring for
LinearCombionation/qml.Hamiltonian. It references using
`qml.Hamiltonian` as an access point.
- The opmath page guiding users in transitioning from legacy operator
arithmetic in the docs is updated to reflect the removal
- A little bit of cleaning up to remove the more obvious instances where
we refer to legacy operators in docstrings that should now refer to
opmath operators. Since "tensor" and "Hamiltonian" are common words
outside of class names, it's hard to do a thorough search, but we got a
lot of the the easy-to-find ones here.

**Benefits:**
Clear documentation surrounding the current operator functionality in
the code base after the removal of legacy operator arithmetic.

[sc-77523]

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>