prune linting on `qnode.py` and `execution.py` (#6621)
**Context:**

Over time, the amount of linting rules we've silenced on `qnode.py` and
`execution.py` has grown to be rather large. In some cases, we are still
silencing things that aren't even a problem anymore. In others, we are
silencing complaints that have extremely easy fixes.

**Description of the Change:**

Prune's the list of things we are silencing in `qnode.py` and
`execution.py` to either:

1) Fix the problem
2) Silence pylint as close to the problem as possible to avoid false
negatives

**Benefits:**

Code more in line with pylint recommendations and fewer false negatives
going forward.

**Possible Drawbacks:**

**Related GitHub Issues:**