Tidy classical cotransform part of `TransformProgram` (#6631)
**Context:**

When trying to improve the organization and structure of our workflow,
one major blocker is being able to handle the classical cotransform and
argnums component of the transform program. Right now we can't defer
adding in the device transforms and gradient preprocessing till inside
`qml.execute`, because we have to set the classical component of the
transform program in the qnode.

One option I am considering is storing `qnode`, `args`, and `kwargs` on
the program and lazily computing args and the classical component,
instead of eagerily calculating the `argnums` and
`classical_cotransforms`. This would allow us to setup the full
transform program inside `qml.execute`, instead of the `QNode`.

The main blocker to making this change was understanding what the
classical component was actually doing. While figuring that out, I
rewrote some of the code to reduce nesting and reduce the amount going
on in each individual component. Hopefully this makes it a little easier
to understand.

**Description of the Change:**

Reduces the levels of nesting in
`TransformProgram._set_all_classical_jacobians` to make it a little
easier to understand.

**Benefits:**

In the future, it will be easier to move components around without
breaking things.

**Possible Drawbacks:**

Maybe an unnecessary waste of time.

**Related GitHub Issues:**